+++ libraries/chef_pyenv_script_helpers.rb
@@ -1,5 +1,5 @@
-class Chef
-  module Pyenv
+class PyEnv
+  module Cookbook
     module ScriptHelpers
       require 'chef/mixin/shell_out'