+++ resources/global.rb
@@ -1,5 +1,5 @@
 unified_mode true
-include Chef::Pyenv::ScriptHelpers
+include PyEnv::Cookbook::ScriptHelpers
 
 property :pyenv_version,
           String,
@@ -15,14 +15,14 @@ property :prefix,
 action :create do
   pyenv_script 'globals' do
     code "pyenv global #{new_resource.pyenv_version}"
-    user new_resource.user if new_resource.user
+    user new_resource.user
     action :run
     not_if { current_global_version_correct? }
   end
 end
 
 action_class do
-  include Chef::Pyenv::ScriptHelpers
+  include PyEnv::Cookbook::ScriptHelpers
 
   def current_global_version_correct?
     current_global_version == new_resource.pyenv_version