+++ test/fixtures/cookbooks/test/recipes/system_install.rb
@@ -1,17 +1,21 @@
-# frozen_string_literal: true
+# Install Python system wide
+# and make it the global default
 
 version = '3.7.7'
-venv_root = '/opt/venv_test'
+local_folder = '/opt/pyenv_test'
 
-cookbook_file '/tmp/requirements.txt' do
-  source 'requirements.txt'
-  mode '0644'
-end
+directory local_folder
 
-# Install pyenv globally
-pyenv_system_install 'system'
+pyenv_install 'system'
+
+# pyenv_rehash 'for-new-python'
 
 pyenv_python version
+
+pyenv_local version do
+  path local_folder
+  user 'root'
+end
 
 pyenv_global version
 
@@ -21,28 +25,4 @@ end
 
 pyenv_pip 'requests' do
   version '2.18.3'
-end
-
-pyenv_pip 'virtualenv' do
-  version '16.2.0'
-end
-
-pyenv_script 'create virtualenv' do
-  code "virtualenv #{venv_root}"
-end
-
-pyenv_pip '/tmp/requirements.txt' do
-  virtualenv venv_root
-  requirement true
-end
-
-pyenv_pip 'urllib3' do
-  virtualenv venv_root
-  action :upgrade
-  version '1.25.11'
-end
-
-pyenv_pip 'requests' do
-  virtualenv venv_root
-  action :uninstall
 end