+++ test/fixtures/cookbooks/test/recipes/user_install.rb
@@ -1,5 +1,7 @@
-# Install pyenv and makes it avilable to the selected user
-version   = '3.6.1'
+# Install pyenv
+# and make it avilable to the selected user
+
+version   = '3.7.7'
 user      = 'vagrant'
 venv_root = "/home/#{user}/venv_test"
 
@@ -10,7 +12,9 @@ cookbook_file '/tmp/requirements.txt' do
   mode '0644'
 end
 
-pyenv_user_install user
+pyenv_install 'user' do
+  user user
+end
 
 pyenv_python version do
   user user